requirements.txt
streamlit
numpy==1.23.2
pandas==1.4.2
Pillow==9.4
notebook==6.5.2
pytest==7.1.3
pytest-mock==3.8.2